third_party/blink/web_tests/platform/generic/fast/dom/webtiming-expected.txt

This test checks that all of the Web Timing attributes are available and have reasonable values in the right order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timing.navigationStart is >= oneHourAgoUTC
PASS timing.redirectStart is 0
PASS timing.redirectEnd is 0
PASS navigation.redirectCount is 0
PASS timing.fetchStart is >= timing.navigationStart
PASS timing.domainLookupStart is >= timing.fetchStart
PASS timing.domainLookupEnd is >= timing.domainLookupStart
PASS timing.connectStart is >= timing.domainLookupEnd
PASS timing.connectEnd is >= timing.connectStart
PASS timing.secureConnectionStart is 0
PASS timing.requestStart is >= timing.connectEnd
PASS timing.responseStart is >= timing.requestStart
PASS timing.domLoading is >= timing.fetchStart
PASS timing.domInteractive is 0
PASS timing.domContentLoadedEventStart is 0
PASS timing.domContentLoadedEventEnd is 0
PASS timing.domComplete is 0
PASS timing.loadEventStart is 0
PASS timing.loadEventEnd is 0
PASS timing.navigationStart is >= oneHourAgoUTC
PASS timing.redirectStart is 0
PASS timing.redirectEnd is 0
PASS navigation.redirectCount is 0
PASS timing.fetchStart is >= timing.navigationStart
PASS timing.domainLookupStart is >= timing.fetchStart
PASS timing.domainLookupEnd is >= timing.domainLookupStart
PASS timing.connectStart is >= timing.domainLookupEnd
PASS timing.connectEnd is >= timing.connectStart
PASS timing.secureConnectionStart is 0
PASS timing.requestStart is >= timing.connectEnd
PASS timing.responseStart is >= timing.requestStart
PASS timing.domLoading is >= timing.fetchStart
PASS timing.domInteractive is >= timing.domLoading
PASS timing.domContentLoadedEventStart is 0
PASS timing.domContentLoadedEventEnd is 0
PASS timing.domComplete is 0
PASS timing.loadEventStart is 0
PASS timing.loadEventEnd is 0
PASS timing.navigationStart is >= oneHourAgoUTC
PASS timing.redirectStart is 0
PASS timing.redirectEnd is 0
PASS navigation.redirectCount is 0
PASS timing.fetchStart is >= timing.navigationStart
PASS timing.domainLookupStart is >= timing.fetchStart
PASS timing.domainLookupEnd is >= timing.domainLookupStart
PASS timing.connectStart is >= timing.domainLookupEnd
PASS timing.connectEnd is >= timing.connectStart
PASS timing.secureConnectionStart is 0
PASS timing.requestStart is >= timing.connectEnd
PASS timing.responseStart is >= timing.requestStart
PASS timing.domLoading is >= timing.fetchStart
PASS timing.domInteractive is >= timing.domLoading
PASS timing.domContentLoadedEventStart is >= timing.domInteractive
PASS timing.domContentLoadedEventEnd is 0
PASS timing.domComplete is 0
PASS timing.loadEventStart is 0
PASS timing.loadEventEnd is 0
PASS timing.navigationStart is >= oneHourAgoUTC
PASS timing.redirectStart is 0
PASS timing.redirectEnd is 0
PASS navigation.redirectCount is 0
PASS timing.fetchStart is >= timing.navigationStart
PASS timing.domainLookupStart is >= timing.fetchStart
PASS timing.domainLookupEnd is >= timing.domainLookupStart
PASS timing.connectStart is >= timing.domainLookupEnd
PASS timing.connectEnd is >= timing.connectStart
PASS timing.secureConnectionStart is 0
PASS timing.requestStart is >= timing.connectEnd
PASS timing.responseStart is >= timing.requestStart
PASS timing.domLoading is >= timing.fetchStart
PASS timing.domInteractive is >= timing.responseEnd
PASS timing.domContentLoadedEventStart is >= timing.domInteractive
PASS timing.domContentLoadedEventEnd is >= timing.domContentLoadedEventStart
PASS timing.domComplete is 0
PASS timing.loadEventStart is 0
PASS timing.loadEventEnd is 0
PASS timing.navigationStart is >= oneHourAgoUTC
PASS timing.redirectStart is 0
PASS timing.redirectEnd is 0
PASS navigation.redirectCount is 0
PASS timing.fetchStart is >= timing.navigationStart
PASS timing.domainLookupStart is >= timing.fetchStart
PASS timing.domainLookupEnd is >= timing.domainLookupStart
PASS timing.connectStart is >= timing.domainLookupEnd
PASS timing.connectEnd is >= timing.connectStart
PASS timing.secureConnectionStart is 0
PASS timing.requestStart is >= timing.connectEnd
PASS timing.responseStart is >= timing.requestStart
PASS timing.responseEnd is >= timing.responseStart
PASS timing.domLoading is >= timing.fetchStart
PASS timing.domInteractive is >= timing.responseEnd
PASS timing.domContentLoadedEventStart is >= timing.domInteractive
PASS timing.domContentLoadedEventEnd is >= timing.domContentLoadedEventStart
PASS timing.domComplete is >= timing.domContentLoadedEventEnd
PASS timing.loadEventStart is >= timing.responseEnd
PASS timing.loadEventEnd is 0
PASS timing.navigationStart is >= oneHourAgoUTC
PASS timing.redirectStart is 0
PASS timing.redirectEnd is 0
PASS navigation.redirectCount is 0
PASS timing.fetchStart is >= timing.navigationStart
PASS timing.domainLookupStart is >= timing.fetchStart
PASS timing.domainLookupEnd is >= timing.domainLookupStart
PASS timing.connectStart is >= timing.domainLookupEnd
PASS timing.connectEnd is >= timing.connectStart
PASS timing.secureConnectionStart is 0
PASS timing.requestStart is >= timing.connectEnd
PASS timing.responseStart is >= timing.requestStart
PASS timing.responseEnd is >= timing.responseStart
PASS timing.domLoading is >= timing.fetchStart
PASS timing.domInteractive is >= timing.responseEnd
PASS timing.domContentLoadedEventStart is >= timing.domInteractive
PASS timing.domContentLoadedEventEnd is >= timing.domContentLoadedEventStart
PASS timing.domComplete is >= timing.domContentLoadedEventEnd
PASS timing.loadEventStart is >= timing.responseEnd
PASS timing.loadEventEnd is >= timing.loadEventStart + 50
PASS successfullyParsed is true

TEST COMPLETE
